refactor: SendMailEditor를 없애고 MailEditor 사용
