+++ src/pages/SendMail/components/MailEditorSection/EditorSection.tsx
@@ -1,5 +1,6 @@
-import { SendMailEditor } from '@/pages/SendMail/components/MailEditorSection/SendMailEditor';
 import { Part } from '@/query/part/schema';
+
+import { MailEditor } from '../MailEditor/MailEditor';
 
 interface EditorSectionProps {
   selectedPart: Part | undefined;
@@ -10,7 +11,7 @@ export const EditorSection = ({ selectedPart, selectedTemplateId }: EditorSectio
   return (
     <div className="border-line-basicMedium bg-bg-basicDefault mx-auto flex h-full max-h-[690px] w-full flex-col rounded-xl border">
       {selectedPart && (
-        <SendMailEditor
+        <MailEditor
           key={selectedPart.partId && selectedTemplateId}
           selectedPart={selectedPart}
           selectedTemplateId={selectedTemplateId}
